# bacon_number/actor_swagger.yml
openapi: 3.0.0
info:
  description: This is the description file of the bacon number app
  version: 1.0.0
  title: Bacon Number REST API

servers:
  - url: http://localhost/api/v1
    description: Development server
  - url: http://api.mydomain.com/v1
    description: Example of production server

paths:
  /actor:
    get:
      summary: "Returns the list of actors"
      description: "Retrieve and return the list of actors"
      operationId: bacon_number.actor.show_all
      parameters:
        - name: offset
          description: The number of items to skip before starting to collect the result set
          in: query
          schema:
            type: integer
        - name: limit
          description: The numbers of items to return
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Successful retrieved list
          content:
            application/json:
              schema:
                items:
                  properties:
                    name:
                      type: string
        default:
          description: Unexpected error

  /actor/{actorName}/bacon_number:
    get:
      summary: "Returns the bacon number of an actor"
      description: "Calculates the bacon number of an actor using a graph of connections"
      operationId: bacon_number.actor.bacon_number
      parameters:
        - in: path
          name: actorName
          schema:
            type: string
          required: true
          description: Name of the actor to calculate the bacon number
      responses:
        200:
          description: "Successful calculated bacon number"
          content:
            application/json:
              schema:
                properties:
                  degrees:
                    type: "int"
        400:
          description: The specified actor name is invalid.
        404:
          description: A actor with the specified name was not found.
        default:
          description: Unexpected error